pyproject.toml
[tool.poetry]
name = "gemini_bot"
version = "0.1.0"
description = "A gemini bot/cli for executing and evaluating one or more trading strategies"
authors = ["pk8189 <pmkelly4444@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
gemini_python = "^0.2.0"
python-dotenv = "^0.10.3"
pandas = "^1.0.1"

[tool.poetry.dev-dependencies]
pylint = "^2.4"

[tool.poetry.scripts]
gemini_bot = "gemini_bot.cli.__main__:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
